requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
ultralytics==8.3.170
ultralytics-thop==2.0.14
torch==2.5.1
torchvision>=0.15.0
pytest>=8.0.0
pandas>=2.0.0
flask>=3.0.0
python-dotenv==1.0.0
easyocr==1.7.2
firebase-admin==6.9.0
flask-marshmallow==1.3.0
tqdm==4.67.1
scipy>=1.10.0
runpod==1.7.12
scikit-learn==1.7.1
networkx==3.5
shapely==2.1.1
tensorflow==2.18.1
gunicorn>=21.2.0
lap>=0.5.12